requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
RPi.GPIO==0.7.1
adafruit-circuitpython-ahtx0>=1.0.0
flask>=3.0.0
